requirements.txt
Flask==2.3.2
tensorflow==2.12.0
numpy==1.24.3
Pillow==9.5.0
flask-cors==3.0.10
